backend/requirements.txt
# Core Framework
Django>=5.0.0
django-ninja>=1.0.1
django-cors-headers>=4.3.1
django-filter>=23.5
django-storages>=1.14.2

# Database
SQLAlchemy>=2.0.25
psycopg2-binary>=2.9.9  # PostgreSQL adapter
alembic>=1.13.1  # Database migrations
asyncpg>=0.29.0  # Async PostgreSQL support
dj-database-url>=2.1.0  # Database URL parsing

# API & Validation
pydantic>=2.6.1
pydantic-settings>=2.1.0
openapi-schema-pydantic>=1.2.4
drf-spectacular>=0.27.0  # OpenAPI/Swagger documentation

# Authentication & Security
PyJWT>=2.8.0
bcrypt>=4.1.2
python-jose>=3.3.0  # Additional JWT support
django-ratelimit>=4.1.0  # Rate limiting
django-axes>=6.3.0  # Security monitoring

# Container Management
docker>=7.0.0  # Docker API client

# Testing
pytest>=8.0.0
pytest-django>=4.8.0
pytest-asyncio>=0.23.5
pytest-cov>=4.1.0
httpx>=0.26.0
faker>=22.6.0  # Test data generation
factory-boy>=3.3.0  # Test factories

# Logging & Monitoring
structlog>=24.1.0
python-json-logger>=2.0.7
sentry-sdk>=1.39.0
prometheus-client>=0.19.0

# Utilities
python-dotenv>=1.0.1
tenacity>=8.2.3  # Retry logic
cachetools>=5.3.2
redis>=5.0.1  # Caching and rate limiting
celery>=5.3.6  # Async task processing
flower>=2.0.1  # Celery monitoring

# Development Tools
black>=24.1.1  # Code formatting
isort>=5.13.2  # Import sorting
flake8>=7.0.0  # Linting
mypy>=1.8.0  # Type checking
pre-commit>=3.6.0  # Git hooks
golang

# Documentation
mkdocs>=1.5.3
mkdocs-material>=9.5.3
mkdocstrings>=0.24.0

# Security Tools
bandit>=1.7.7  # Security linter
safety>=2.3.5  # Dependency security checker